+++ packages/controls/js/libs/index.js
@@ -125,6 +125,14 @@ export { default as PanelBodyControl } from './panel-body-control';
 export { default as CustomPropertyControl } from './custom-property-control';
 export { default as SearchReplaceControl } from './search-replace-control';
 export { default as NoticeControl } from './notice-control';
+export {
+	Snackbar,
+	SnackbarList,
+	SnackbarNotices,
+	createSnackbarNotice,
+	BLOCKERA_SNACKBAR_CONTEXT,
+	DefaultSnackbarIcon,
+} from './snackbar';
 export { default as PositionButtonControl } from './position-button';
 export { default as DividerControl } from './divider-control';
 export { default as MaskControl } from './mask-control';