+++ packages/controls/js/libs/snackbar/constants.js
@@ -1,0 +1,3 @@
+// @flow
+
+export const BLOCKERA_SNACKBAR_CONTEXT = 'blockera';