+++ packages/controls/js/libs/snackbar/default-icon.js
@@ -1,0 +1,22 @@
+// @flow
+
+/**
+ * External dependencies
+ */
+import type { MixedElement } from 'react';
+
+/**
+ * Blockera dependencies
+ */
+import { Icon } from '@blockera/icons';
+
+export function DefaultSnackbarIcon(): MixedElement {
+	return (
+		<Icon
+			library={'blockera'}
+			icon={'blockera'}
+			iconSize={18}
+			className={'blockera-component-snackbar__brand-icon'}
+		/>
+	);
+}