+++ packages/controls/js/libs/snackbar/index.js
@@ -1,0 +1,8 @@
+// @flow
+
+export { default as Snackbar } from './snackbar';
+export { default as SnackbarList } from './snackbar-list';
+export { default as SnackbarNotices } from './snackbar-notices';
+export { createSnackbarNotice } from './create-snackbar-notice';
+export { BLOCKERA_SNACKBAR_CONTEXT } from './constants';
+export { DefaultSnackbarIcon } from './default-icon';